// src/config/constants/types.ts
export enum QuoteToken {
  "BNB" = "BNB",
  "CAKE" = "CAKE",
  "SYRUP" = "SYRUP",
  "BUSD" = "BUSD",
  "TWT" = "TWT",
  "UST" = "UST",
  "USDC" = "USDC",
  "USDT" = "USDT",
  'RASTA' = 'RASTA',
  'MRASTA' = 'MRASTA',
  'ETH' = 'ETH',
  'COMP' = 'COMP',
  'DOT' = 'DOT',
  'CNS' = 'CNS',
  'DFL' = 'DFL',
}

export enum PoolCategory {
  "COMMUNITY" = "Community",
  "CORE" = "Core",
  "BINANCE" = "Binance", // Pools using native BNB behave differently than pools using a token
}

export interface Address {
  97?: string;
  56: string;
}
// ...
export interface GrandPoolRewardToken {
  tokenSymbol: string;
  tokenAddress: Address;
  tokenPrice: number;
  rewardPerBlock: number;
}
export interface GrandFarmConfig {
  pid: number;
  pidv1?: number;
  lpSymbol: string;
  lpType: string;
  lpAddresses: Address;
  poolAddress: Address;
  tokenIcon?: string;
  tokenSymbol: string;
  tokenAddresses: Address;
  rewardTokenSymbol?: string;
  rewardTokenAddresses?: Address;
  rewardTokens?: Array<GrandPoolRewardToken>;
  quoteTokenSymbol: QuoteToken;
  quoteTokenAdresses: Address;
  multiplier?: string;
  isTokenOnly?: boolean;
  isCommunity?: boolean;
  risk: number;
  dual?: {
    rewardPerBlock: number;
    earnLabel: string;
    endBlock: number;
  };
  title: string;
  tokenDecimal?: number;
  tokenPrice?: number;
  rewardTokenPrice?: number;
  quoteTokenPrice?: number;
  newPool?: number;
  duration?: number;
  isWhalePool?: boolean;
  isShrimpPool?: boolean;
  isBoosterPool?: boolean;
  isDualFarm?: boolean;
  depositLink?: string;
  projectLink?: string;
  active: boolean;
  isLPToken?: boolean;
  isHintVisible?: boolean;
  minStaking?: number;
  showBackground?: boolean;
  penaltyFee?: number;
  minFirstDeposit?: number;
  maxFirstDeposit?: number;
  reward1?: number;
  poolStartTime?: number;
  reward2?: number;
  cgProjectID?: string;
  closed?: boolean;
  revert?: boolean;
  lpTokenDecimal?: number;
}